mindlancer-project/Backend (Django) - Server & Database/proposals/views.py
from rest_framework import generics
from .models import Proposal
from .serializers import ProposalSerializer

class ProposalListCreateView(generics.ListCreateAPIView):
    queryset = Proposal.objects.all()
    serializer_class = ProposalSerializer

class ProposalDetailView(generics.RetrieveUpdateDestroyAPIView):
    queryset = Proposal.objects.all()
    serializer_class = ProposalSerializer
