mindlancer-project/Backend (Django) - Server & Database/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.2.0
djangorestframework-simplejwt==5.3.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9  # For PostgreSQL
dj-database-url==2.0.0
django-allauth==0.58.2  # Social auth
django-filter==23.5  # For job search
gunicorn==21.2.0  # For deployment
whitenoise==6.6.0  # Static files
python-dateutil==2.8.2  # Date handling